requirements_cpu.txt
# CPU优化版依赖 - 适用于ZEABUR部署
# 基础 Flask 依赖
Flask==3.1.2
Flask-CORS==6.0.1
jsonschema==4.25.1
python-dotenv==1.1.1

# CPU优化的语音识别 (比标准whisper快4-5倍)
faster-whisper==1.0.2

# 音频视频处理
ffmpeg-python==0.2.0

# 基础工具
requests==2.31.0
psutil==5.9.0
APScheduler==3.11.0
srt==3.5.3
numpy==1.26.0

# 图像处理 (轻量版)
Pillow==10.1.0

# 时间处理
python-dateutil==2.8.2

# 验证和工具
attrs==23.1.0
referencing==0.30.2
rpds-py==0.12.0
typing-extensions==4.8.0

# 音频处理工具
librosa==0.10.1  # 音频分析
soundfile==0.12.1  # 音频文件读写



